+++ client/src/stories/Dogs.stories.js
@@ -1,0 +1,14 @@
+import Dogs from "../components/search-page/Dogs";
+
+export default {
+  title: "Search-Page/Dogs",
+  component: Dogs,
+};
+
+const Template = (args) => <Dogs {...args} />;
+
+export const Friend = Template.bind({});
+Friend.args = {
+  name: "Steve",
+  imgSrc: "blah.jpg",
+};
