+++ client/src/stories/SingleDog.stories.js
@@ -1,0 +1,18 @@
+import SingleDog from "../components/single-dog/SingleDog";
+
+export default {
+  title: "single-dog/SingleDog",
+  component: SingleDog,
+};
+
+const Template = (args) => <SingleDog {...args} />;
+
+export const Dog = Template.bind({});
+Dog.args = {
+  stats: ["loves scratches", "hates cats"],
+  imgSrc: "blah.jpg",
+  bio: "Hi, I love walks in the morning and going to the beach",
+  rating: Number,
+  review: "not the nicest of dogs",
+  isStarred: false,
+};
